+++ src/vs/workbench/contrib/chat/browser/aiCustomization/aiCustomizationTreeView.contribution.ts
@@ -3,17 +3,54 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { localize, localize2 } from '../../../../nls.js';
-import { Action2, MenuRegistry, registerAction2 } from '../../../../platform/actions/common/actions.js';
-import { ContextKeyExpr } from '../../../../platform/contextkey/common/contextkey.js';
-import { ServicesAccessor } from '../../../../platform/instantiation/common/instantiation.js';
-import { AICustomizationItemMenuId } from './aiCustomizationTreeView.js';
-import { AICustomizationItemTypeContextKey } from './aiCustomizationTreeViewViews.js';
-import { PromptsType } from '../../../../workbench/contrib/chat/common/promptSyntax/promptTypes.js';
-import { Codicon } from '../../../../base/common/codicons.js';
-import { ICommandService } from '../../../../platform/commands/common/commands.js';
-import { URI } from '../../../../base/common/uri.js';
-import { IEditorService } from '../../../../workbench/services/editor/common/editorService.js';
+import { localize, localize2 } from '../../../../../nls.js';
+import { Action2, MenuRegistry, registerAction2 } from '../../../../../platform/actions/common/actions.js';
+import { ContextKeyExpr } from '../../../../../platform/contextkey/common/contextkey.js';
+import { SyncDescriptor } from '../../../../../platform/instantiation/common/descriptors.js';
+import { ServicesAccessor } from '../../../../../platform/instantiation/common/instantiation.js';
+import { Registry } from '../../../../../platform/registry/common/platform.js';
+import { AICustomizationItemMenuId, AI_CUSTOMIZATION_VIEW_ID } from './aiCustomizationTreeView.js';
+import { AICustomizationItemTypeContextKey, AICustomizationViewPane } from './aiCustomizationTreeViewViews.js';
+import { AI_CUSTOMIZATION_OVERVIEW_VIEW_ID, AICustomizationOverviewView } from './aiCustomizationOverviewView.js';
+import { PromptsType } from '../../common/promptSyntax/promptTypes.js';
+import { Codicon } from '../../../../../base/common/codicons.js';
+import { ICommandService } from '../../../../../platform/commands/common/commands.js';
+import { URI } from '../../../../../base/common/uri.js';
+import { IEditorService } from '../../../../services/editor/common/editorService.js';
+import { ChatContextKeys } from '../../common/actions/chatContextKeys.js';
+import { ChatViewContainerId } from '../chat.js';
+import { Extensions as ViewExtensions, IViewContainersRegistry, IViewsRegistry } from '../../../../common/views.js';
+
+//#region View Registration
+
+const viewContainerRegistry = Registry.as<IViewContainersRegistry>(ViewExtensions.ViewContainersRegistry);
+const viewsRegistry = Registry.as<IViewsRegistry>(ViewExtensions.ViewsRegistry);
+const chatViewContainer = viewContainerRegistry.get(ChatViewContainerId);
+
+if (chatViewContainer) {
+	viewsRegistry.registerViews([
+		{
+			id: AI_CUSTOMIZATION_VIEW_ID,
+			name: localize2('aiCustomization.view', "AI Customizations"),
+			ctorDescriptor: new SyncDescriptor(AICustomizationViewPane),
+			canToggleVisibility: true,
+			canMoveView: true,
+			when: ChatContextKeys.enabled,
+			order: 100,
+		},
+		{
+			id: AI_CUSTOMIZATION_OVERVIEW_VIEW_ID,
+			name: localize2('aiCustomizationOverview.view', "AI Customizations Overview"),
+			ctorDescriptor: new SyncDescriptor(AICustomizationOverviewView),
+			canToggleVisibility: true,
+			canMoveView: true,
+			when: ChatContextKeys.enabled,
+			order: 101,
+		},
+	], chatViewContainer);
+}
+
+//#endregion
 
 //#region Utilities
 
@@ -92,7 +129,5 @@ MenuRegistry.appendMenuItem(AICustomizationItemMenuId, {
 	order: 1,
 	when: ContextKeyExpr.equals(AICustomizationItemTypeContextKey.key, PromptsType.prompt),
 });
-
-//#endregion
 
 //#endregion
