+++ src/vs/workbench/contrib/chat/browser/aiCustomization/media/aiCustomizationManagement.css
