+++ src/vs/workbench/contrib/chat/browser/aiCustomization/media/aiCustomizationTreeView.css
